Create helper functions in FlowControl_IType
Basically what we do in the case of predicting and recovering from
branches is the same regardless of the specific branch instruction, so
two helper classes are added to FlowControl_IType (all branch
instructions are children of this class).
makePrediction uses the predictor to make a prediction and change the
program counter accordingly.
respondToCondition responds to the resolved condition of the branch
(true or false), updates the relevant predictor, and triggers an
exception if the prediction was wrong.

This is currently only used in BEQZ, but should be easy to add to any
other branch instruction.