Bug fix in local predictor table, add other branches
Basically the predictor fetched in IF was not necessarily the same
predictor used in ID. This leads to the situation where a prediction is
correctly made, but the software acts as if the prediction was in error
and can sometimes fetch a completely unrelated instruction from a
previous branch.
This was resolved by rewriting getLocalPrediction to directly use the
instruction, not the program counter.

Also implements the dynamic branch prediction stuff for the other branch
instructions.